Some miscellaneous changes
* WIP file for including Cakelisp at runtime
* makeDirectory now returns success/failure
* Use size_t for loop iteration to fix compilation errors
* Add token-splice keywords to Emacs config

+++ src/RunProcess.cpp
@@ -414,7 +414,7 @@ void waitForAllProcessesClosed(SubprocessOnOutputFunc onOutput)
 	if (s_subprocesses.empty())
 		return;
 
-	for (int i = 0; i < s_subprocesses.size(); ++i)
+	for (size_t i = 0; i < s_subprocesses.size(); ++i)
 	{
 		Subprocess* process = &s_subprocesses[i];
 #if defined(UNIX) || defined(MACOS)
